% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_vertical.R
\name{th_vertical}
\alias{th_vertical}
\title{Title: estimating function for three causes vertical model}
\usage{
th_vertical(
  xtime,
  Delta,
  D1,
  D2,
  X,
  Z,
  Bt,
  W,
  Q,
  ibeta,
  igamma1,
  igamma2,
  ikappa1,
  ikappa2,
  iU,
  izetah,
  izetal,
  irho,
  h1,
  h2,
  itheta,
  lr = 1,
  mite = 20L,
  miter = 20L,
  eps = 0.001
)
}
\arguments{
\item{xtime:observed}{time}

\item{Delta:censoring}{indicator}

\item{D1:cause}{2 indicator}

\item{D2:cause}{3 inicator}

\item{X:covariates}{of total hazards model}

\item{Z:covariates}{of relative hazards model}

\item{Bt:b-spline}{basis for estimating function in relative hazards model}

\item{W:threshold}{variable}

\item{Q:grouping}{indicator}

\item{ibeta:initial}{values of covariates cofficients in total hazards model}

\item{igamma1:initial}{values of covariates cofficients in cause 2 of relative hazards model}

\item{igamma2:initial}{values of covariates cofficients in cause 3 of relative hazards model}

\item{ikappa1:initial}{values of b-spline cofficients  in cause 2 of relative hazards model}

\item{ikappa2:initial}{values of b-spline cofficients  in cause 3 of relative hazards model}

\item{iU:initial}{values of frailty}

\item{izetah:initial}{value of threshold in total hazards model}

\item{izetal:initial}{value of threshold in relative hazards model}

\item{irho:initial}{value of frailty cofficients in relative hazards model}

\item{h1:bandwidth}{of  the S-shaped approximation function in threshold estimation of total hazards model}

\item{h2:bandwidth}{of  the S-shaped approximation function in threshold estimation of relative hazards model}

\item{itheta:initial}{value of variance estimation of frailties}

\item{lr:step}{size reduction multiple}

\item{mite:The}{maximum number of iterations for coefficient estimation and frailties estimation}

\item{miter:The}{maximum number of iterations for variance estimation  and coefficient estimation of frailties}

\item{eps:}{}
}
\description{
Title: estimating function for three causes vertical model
}
