% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_vertical.R
\name{tw_vertical}
\alias{tw_vertical}
\title{Title}
\usage{
tw_vertical(
  xtime,
  Delta,
  D,
  X,
  Z,
  Bt,
  W,
  Q,
  ibeta,
  igamma,
  ikappa,
  iU,
  izetah,
  izetal,
  irho,
  h1,
  h2,
  itheta,
  lr = 1,
  mite = 20L,
  miter = 20L,
  eps = 0.001,
  rho_bound = 1,
  zeta_lr_pi = 1
)
}
\arguments{
\item{xtime}{}

\item{Delta}{}

\item{D}{}

\item{X}{}

\item{Z}{}

\item{Bt}{}

\item{W}{}

\item{Q}{}

\item{ibeta}{}

\item{igamma}{}

\item{ikappa}{}

\item{iU}{}

\item{izetah}{}

\item{izetal}{}

\item{irho}{}

\item{h1}{}

\item{h2}{}

\item{itheta}{}

\item{lr}{}

\item{mite}{}

\item{miter}{}

\item{eps}{}

\item{rho_bound}{Positive upper bound for the absolute frailty loading.}

\item{zeta_lr_pi}{Learning-rate multiplier for the relative-threshold update.}
}
\description{
Title
}
