\name{VerticalCR-package}
\alias{VerticalCR-package}
\alias{VerticalCR}
\docType{package}
\title{
\packageTitle{VerticalCR}
}
\description{
\packageDescription{VerticalCR}
}
\details{
See the documentation for Armadillo, and RcppArmadillo, for more details.
}
\author{
\packageAuthor{VerticalCR}

Maintainer: \packageMaintainer{VerticalCR}
}
\references{
See the documentation for Armadillo, and RcppArmadillo, for more details.
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = VerticalCR)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{VerticalCR}}
%\section{Documentation Index}{\packageIndices{VerticalCR}}

\seealso{
See the documentation for Armadillo, and RcppArmadillo, for more details.
}

