\name{verticaltry-package}
\alias{verticaltry-package}
\alias{verticaltry}
\docType{package}
\title{
\packageTitle{verticaltry}
}
\description{
\packageDescription{verticaltry}
}
\details{
See the documentation for Armadillo, and RcppArmadillo, for more details.
}
\author{
\packageAuthor{verticaltry}

Maintainer: \packageMaintainer{verticaltry}
}
\references{
See the documentation for Armadillo, and RcppArmadillo, for more details.
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = verticaltry)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{verticaltry}}
%\section{Documentation Index}{\packageIndices{verticaltry}}

\seealso{
See the documentation for Armadillo, and RcppArmadillo, for more details.
}

